% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSynData.R
\name{AddAllScripts}
\alias{AddAllScripts}
\title{Create scripts to run SignatureAnalyzer in all subdirectories
of \code{OutDir()}.}
\usage{
AddAllScripts(maxK = 30, top.level.dir = NULL)
}
\arguments{
\item{maxK}{The \code{maxK} argument for SignatureAnalyzer.}

\item{top.level.dir}{Add the scripts to sub-directories
\code{sa.sa.96}, \code{sp.sp}, etc. of this
directory.}
}
\description{
Create scripts to run SignatureAnalyzer in all subdirectories
of \code{OutDir()}.
}
\keyword{internal}
