% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSynData.R
\name{AddNoise}
\alias{AddNoise}
\title{Add Poisson noise to each mutation type in each signature in an epxosure.}
\usage{
AddNoise(input.exposure, signatures)
}
\arguments{
\item{input.exposure}{The exposures to which to add noise; a numeric matrix
or data frame in which the rows are signatures and the columns are
samples. Each cell indicates the number of mutations due to a particular
signature in a particular sample.}

\item{signatures}{The signatures in the exposure; the column names
of \code{signatures} have to include all row names in
\coce{input.exposure}; can be an \code{\link[ICAMS]{ICAMS}}
catalog or a numerical matrix or data frame.}
}
\value{
A list with the elements \describe{
\item{expsoures}{The numbers of mutations due to each signature
   after adding noise}
\item{spectra}{The spectra based on the noisy signature exposures.}
}
}
\description{
Add Poisson noise to each mutation type in each signature in an epxosure.
}
