% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddAllScripts.R
\name{AddScript}
\alias{AddScript}
\title{Add an R script to run SignatureAnalyzer a particular subdirectory.}
\usage{
AddScript(maxK, slice, dir.name)
}
\arguments{
\item{maxK}{The \code{maxK} argument for SignatureAnalyzers.}

\item{slice}{Which subdirectory to put the script into.}

\item{dir.name}{The name of the subdirectory.}
}
\description{
Add an R script to run SignatureAnalyzer a particular subdirectory.
}
\keyword{internal}
