% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateAbstract.R
\name{Create.3.4.40.Abstract}
\alias{Create.3.4.40.Abstract}
\title{This function supersedes the second part of \code{data-raw/Create.3.5.40.Rmd}.}
\usage{
Create.3.4.40.Abstract(
  seed = 44,
  overwrite = TRUE,
  regress.dir = "data-raw/long.test.regression.data/syn.3.5.40.abst/"
)
}
\description{
With default arguments, this function generates the same results as
the second half of \code{data-raw/Create.3.5.40.Rmd}.
}
\keyword{internal}
