% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateAbstract_Nat.R
\name{Create.3.5.40.Abstract}
\alias{Create.3.5.40.Abstract}
\title{Create synthetic spectra based on SBS3 SBS5 SBS40}
\usage{
Create.3.5.40.Abstract(
  seed = 44,
  overwrite = TRUE,
  regress.dir = "data-raw/long.test.regression.data/syn.3.5.40.abst/",
  num.syn.tumors = 1000,
  top.level.dir = paste0("../syn.3.5.40.abst.", seed),
  unlink = FALSE
)
}
\arguments{
\item{seed}{A random seed to use.}

\item{overwrite}{If \code{TRUE}, overwrite existing directories / files.}

\item{regress.dir}{If not \code{NULL}, compare the result to
the contents of this directory with a \code{diff}.}

\item{num.syn.tumors}{The number of tumors to create \strong{for each cancer
type} in \code{cancer.types}.}

\item{top.level.dir}{The directory in which to put the output; will
be created if necessary.}

\item{unlink}{If \code{TRUE} and \code{!is.null(regress.dir)}, then
unlink the result directory if there are no differences.}
}
\description{
This function generates synthetic spectra with mutation loads of SBS3
(signature prevalent in ovarian adenocarcinoma),
 SBS5 and SBS40 (signatures prevalent in renal cell carcinoma). This
dataset challenges the computational approaches as these three signatures
are "flat" signatures hard to be extracted accurately.
}
\details{
This function supersedes the second part of \code{data-raw/Create.3.5.40.Rmd}
in GitHub repository \code{steverozen/SynSig}. With default arguments, this function generates the same results as
the second half of \code{data-raw/Create.3.5.40.Rmd}.

Data set generated by this function can be found at Synapse with Synapse ID:
\href{https://www.synapse.org/#!Synapse:syn18500215}{syn18500215}.
}
