% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create.syn.many.types.R
\name{CreateMixedTumorTypeSyntheticData}
\alias{CreateMixedTumorTypeSyntheticData}
\title{Create a test data set based on >= 1 tumor types.}
\usage{
CreateMixedTumorTypeSyntheticData(top.level.dir, cancer.type.strings,
  num.syn.tumors, overwrite = FALSE, sa.exp = sa.all.real.exposures,
  sp.exp = sp.all.real.exposures, verbose = 0,
  bladder.regress.hack = FALSE)
}
\arguments{
\item{top.level.dir}{Path to top level of directory structure to be created.}

\item{cancer.type.strings}{Search the PCAWG data for tumors matching
these strings. Each string should identify one tumor type, for
some definition of tumor type. Probably the tumors in each type
should be non-overlapping, but the code does not enforce this and
does not care.}

\item{num.syn.tumors}{Number of synthetic tumors to create
for each cancer type.}

\item{overwrite}{If TRUE, overwrite existing directories / files.}

\item{sa.exp}{SignatureAnalyzer exposures from which to select cancer types
specified by \code{cancer.type.strings}. In the
column names of \code{sa.exp} the cancer type string
should be separated from the sample identifier by two colons
(::).}

\item{sp.exp}{SigProfiler exposures from which to select cancer types
specified by \code{cancer.type.strings}. In the
column names of \code{sp.exp} the cancer type string
should be separated from the sample identifier by two colons
(::).}

\item{verbose}{If > 0, cat various messages.}

\item{bladder.regress.hack}{For use by \code{\link{BladderSkin1000}}.
Forces use of non-hyper-mutated exposures for bladder-TCC even if
\code{sa.exp} and \code{sp.exp} include hyper-mutated exposures.}
}
\description{
Create a test data set based on >= 1 tumor types.
}
