% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateRandomExposures.R
\name{CreateNumSigsPerTumor}
\alias{CreateNumSigsPerTumor}
\title{Create \code{num.exposures} signature counts from a normal distribution with \code{mean} and \code{sd}.}
\usage{
CreateNumSigsPerTumor(
  num.exposures,
  mean.num.sigs.per.tumor,
  sd.num.sigs.per.tumor,
  total.num.sigs
)
}
\arguments{
\item{num.exposures}{Number of exposures to create.}

\item{mean.num.sigs.per.tumor}{Mean of distribution to draw from.}

\item{total.num.sigs}{Number of signatures in the "universe".}

\item{sd}{Standard deviation of distribution to draw from.}
}
\value{
A numeric vector, each element of which the number
of signatures in the corresponding tumor (which still remains
to be created.)
}
\description{
Discard tumors with signature count = 0 or
signature count > \code{total.num.sigs}.
}
\keyword{internal}
