% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateRandomExposures.R
\name{CreateRandomExposures}
\alias{CreateRandomExposures}
\title{Create synthetic exposure data.}
\usage{
CreateRandomExposures(num.exposures, mean.num.sigs.per.tumor,
  sd.num.sigs.per.tumor, total.num.sigs, per.sig.mean.and.sd,
  sample.name.prefix, sigs, verbose)
}
\description{
Create synthetic exposure data.
}
\keyword{internal}
