% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateRandom.R
\name{CreateRandomSyn}
\alias{CreateRandomSyn}
\title{This is the top-level function to create a set of spectra from random signatures.}
\usage{
CreateRandomSyn(top.level.dir, seed = 1443196,
  regress.dir = "data-raw/long.test.regression.data/syn.30.random.sigs/",
  num.syn.tumors = 1000, overwrite = FALSE, unlink = FALSE,
  verbose = FALSE)
}
\arguments{
\item{top.level.dir}{Directory in which to put all results. It will be
created if necessary.}

\item{seed}{Use default for regression testing.}

\item{regress.dir}{If not \code{NULL} compare the known results in
this directory with the created results in \code{top.level.dir}.}

\item{num.syn.tumors}{Total number of synthetic tumors to create. Use the
default for regression testing.}

\item{overwrite}{If \code{TRUE} overwrite existing files and directories.}

\item{unlink}{If \code{TRUE} unlink the created directory after the
regression test.}

\item{verbose}{If \code{TRUE} print a few informative messages.}
}
\description{
This is the top-level function to create a set of spectra from random signatures.
}
