% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSynSBS1SB5Correlated.R
\name{CreateSBS1SBS5CorrelatedSyntheticData}
\alias{CreateSBS1SBS5CorrelatedSyntheticData}
\title{Function to generate 20 SBS1-SBS5-correlated Synthetic datasets used in testing.}
\usage{
CreateSBS1SBS5CorrelatedSyntheticData(
  top.level.dir = "./",
  regress.dir = NULL,
  overwrite = FALSE,
  add.info = TRUE,
  unlink = FALSE
)
}
\arguments{
\item{top.level.dir}{Top-level-folder to place 20 spectra
datasets generated by this function.
Default: ./ (Current working directory)}

\item{regress.dir}{If not \code{NULL}, compare the result to
the contents of this directory with a \code{diff}.}

\item{overwrite}{Whether to overwrite
(Default: FALSE)}

\item{add.info}{Whether to generate additional information.

You should set it to \code{FALSE} when you want to make a \code{diff}
(i.e. \code{regressdir} is not \code{NULL}). This is because
Additional information may differ on different OS or R sessions,
thus may prevent the dataset from passing the \code{NewDiff4SynDatasets} check.
(Default: TRUE)}

\item{unlink}{Whether to delete temporary dataset folder \code{top.level.dir}.
(Set to TRUE for testing)}
}
\description{
This function is a wrapper around \code{\link{CreateSBS1SBS5CorrelatedSyntheticDataOneDataset}}.
It will use the default parameters to repeat the results.
}
\details{
This function will generate 20 datasets, each with files listed below:

ground.truth.syn.catalog.csv: Generated tumor spectra in
ICAMS SBS96 CSV format.

ground.truth.syn.exposures.csv: Mutation burdens of SBS1 and
SBS5 in generated tumor spectra in ICAMS CSV format.

ground.truth.syn.sigs.csv: Ground-truth SBS1 and SBS5
signatures in ICAMS SBS96 CSV format.

parameters.txt: Parameters used to generate the exposures
and tumor spectra.

scatterplot.pdf: scatterplot illustrating correlation of
exposures of two signatures in generated spectra

seedInUse.txt, RNGInUse.txt: seed and Random Number Generator
used in generation. (For better reproducibility)

sessionInfo.txt: information related to R versions, platforms,
loaded or imported packages, etc. (For better reproducibility)
}
