% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateRandomExpRandomSigs.R
\name{GenerateAllRandomSA}
\alias{GenerateAllRandomSA}
\title{Generate a suite of data with random signatures and exposures.}
\usage{

  GenerateAllRandomSA(top.level.dir = "../SA.like.rand.exp.rand.sigs.2019.08.26",
  overwite = TRUE, verbose = TRUE, num.replicates = 10)
}
\description{
Generate a suite of data with random signatures and exposures.
}
\keyword{(exposures).}
\keyword{Base}
\keyword{SignatureAnalyzer}
\keyword{attributions}
\keyword{internal}
\keyword{on}
\keyword{parameters}
