% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateListOfSigParams.R
\name{GenerateListOfSigParams}
\alias{GenerateListOfSigParams}
\title{Generate a list of signature parameters for different cancer types from real exposure}
\usage{
GenerateListOfSigParams(
  real.exposures,
  cancer.types,
  distribution = NULL,
  verbose = 0,
  sig.params = NULL
)
}
\arguments{
\item{real.exposures}{A matrix of real exposures.}

\item{cancer.types}{A vector of character strings denoting different cancer
types. This function will search \code{real.exposures} for exposures from
tumors matching these strings. See \code{PCAWG7::CancerTypes()} for example.}

\item{distribution}{Probability distribution used to generate synthetic
exposures due to active mutational signatures. Can be \code{neg.binom}
which stands for negative binomial distribution. If \code{NULL} (Default),
then this function uses log normal distribution with base 10.}

\item{verbose}{If > 0 cat various messages.}

\item{sig.params}{Empirical signature parameters generated using real
exposures irrespective of their cancer types. If there
is only one tumor having a signature in a cancer type in \code{real.exposures},
we cannot fit the \code{distribution} to only one data point. Instead, we
will use the empirical parameter \code{size} from \code{sig.params}.
Users can use \code{SynSigGen:::GetSynSigParamsFromExposuresOld} to generate
their own signature parameters. If \code{NULL}(default), this function uses the
PCAWG7 empirical signature parameters. See \code{signature.params} for more details.}
}
\description{
Generate a list of signature parameters for different cancer types from real exposure
}
\note{
This function calls \code{GetSynSigParamsFromExposures}.
}
\examples{

# Generate a list of signature parameters for Indel (ID) using negative binomial distribution
input.sigs.ID <- PCAWG7::signature$genome$ID
real.exposures.ID <- PCAWG7::exposure$PCAWG$ID
cancer.types <- PCAWG7::CancerTypes()[1:5]
sig.params <- SynSigGen::signature.params$ID
ID.sig.params <-
  GenerateListOfSigParams(real.exposures = real.exposures.ID,
                          cancer.types = cancer.types,
                          distribution = "neg.binom",
                          sig.params = sig.params
 )
}
