% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSynData.R
\name{GenerateSynExposureOneSample}
\alias{GenerateSynExposureOneSample}
\title{Using parameters given to generate exposures for synthetic tumors}
\usage{
GenerateSynExposureOneSample(
  tumor,
  sig.interest,
  burden.per.sig,
  sd.per.sig,
  sig.matrix = NULL,
  distribution = NULL,
  sig.params = NULL
)
}
\arguments{
\item{tumor}{Signature presence matrix or exposure matrix for a tumor.
It has only one row, and K (# of signatures) columns.
Value in each column is the presence flag for a mutational signature:
the value can be non-zero(signature is present) or 0(absent).
The name of each column should be the name of a signature.}

\item{sig.interest}{Names of mutational signatures you want to use to
generate exposures. It can be all, or part of signatures in colnames(tumor).}

\item{burden.per.sig}{Mean mutation burden a log10 of the
counts of mutations per megabase.
It has one row, and K columns. Each column name refers to a mutational signature.}

\item{sd.per.sig}{standard deviation of mutation burden.
It has one row, and K columns. Each column name refers to a mutational signature.}

\item{sig.matrix}{Signature matrix to construct synthetic tumors.}

\item{distribution}{Probability distribution used to generate synthetic
exposures due to active mutational signatures. Can be \code{neg.binom}
which stands for negative binomial distribution. If \code{NULL} (Default),
then this function uses log normal distribution with base 10.}

\item{sig.params}{Parameters from \code{\link{GetSynSigParamsFromExposures}}
or another source. Should be
a matrix or data frame with one column for
each signature and the following rows:
\itemize{
\item For log normal distribution,
\describe{
\item{prob}{The proportion of tumors with the signature.}
\item{mean}{The mean(log_10(number of mutations)).}
\item{stdev}{The stdev(log_10(number of mutations)).}
}
\item For negative binomial distribution,
\describe{
\item{prob}{The proportion of tumors with the signature.}
\item{size}{Dispersion parameter.}
\item{mu}{Mean.}
}
}

The rownames need to be the column names of a signature
catalog.}
}
\description{
Using parameters given to generate exposures for synthetic tumors
}
\details{
Determine the intensity of each
mutational signature in a tumor, returning the number of mutations
using the mean mutation burden per signature and the std dev
}
\keyword{internal}
