% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateSyntheticTumors.R
\name{GenerateSyntheticExposuresNew}
\alias{GenerateSyntheticExposuresNew}
\title{Create synthetic exposures based given parameters}
\usage{
GenerateSyntheticExposuresNew(sig.params, num.samples = 10, name = "synthetic")
}
\arguments{
\item{sig.params}{Parameters from \code{\link{GetSynSigParamsFromExposures}}
  or another source. Should be
  a matrix or data frame with one column for
  each signature and the following rows:
\describe{
\item{prob}{The proportion of tumors with the signature.}
\item{mean}{The mean(log_10(number of mutations)).}
\item{stdev}{The stdev(log_10(number of mutations)).}
}
  The rownames need to be the column names of a signature
  catalog.}

\item{num.samples}{Number of samples to generate}

\item{name}{Prefix for sample identifiers in the simulated dataset}
}
\value{
A matrix with the rows being each signature and the columns being
generated samples. Each entry is the count of mutations due to one
signature in one sample.
}
\description{
Create synthetic exposures based given parameters
}
