% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateSyntheticTumors.R
\name{GenerateSyntheticTumors}
\alias{GenerateSyntheticTumors}
\title{Generate synthetic tumors based on real exposures in one or more cancer types}
\usage{
GenerateSyntheticTumors(
  seed,
  dir,
  cancer.types,
  samples.per.cancer.type,
  input.sigs,
  real.exposures,
  sample.prefix.name = "SP.Syn.",
  overwrite = TRUE
)
}
\arguments{
\item{seed}{A random seed to use.}

\item{dir}{The directory in which to put the output; will be created if
necessary.}

\item{cancer.types}{A vector of character strings denoting different cancer
types. This function will search \code{real.exposures} for exposures from
tumors matching these strings. See \code{PCAWG7::CancerTypes()} for example.}

\item{samples.per.cancer.type}{Number of synthetic tumors to create
for each cancer type.}

\item{input.sigs}{A matrix of signatures.}

\item{real.exposures}{A matrix of real exposures.}

\item{sample.prefix.name}{Prefix name to add to the synthetic tumors.}

\item{overwrite}{If TRUE, overwrite existing directories and files.}
}
\value{
A list of three elements that comprise the
synthetic data: \enumerate{
 \item \code{ground.truth.catalog}: Spectra catalog with rows denoting mutation
 types and columns denoting sample names.
 \item \code{ground.truth.signatures}: Signatures active
 in \code{ground.truth.catalog}.
 \item \code{ground.truth.exposures}: Exposures of \code{ground.truth.signatures}
 in \code{ground.truth.catalog}.
}
}
\description{
Generate synthetic tumors based on real exposures in one or more cancer types
}
\examples{

# Generate synthetic tumors for DBS78
input.sigs.DBS78 <- PCAWG7::signature$genome$DBS78
real.exposures.DBS78 <- PCAWG7::exposure$PCAWG$DBS78
cancer.types <- PCAWG7::CancerTypes()[1:5]
DBS78.synthetic.tumors <-
  GenerateSyntheticTumors(seed = 191906,
                          dir = file.path(tempdir(), "DBS78.synthetic.tumors"),
                          cancer.types = cancer.types,
                          samples.per.cancer.type = 30,
                          input.sigs = input.sigs.DBS78,
                          real.exposures = real.exposures.DBS78,
                          sample.prefix.name = "SP.Syn."
  )

# Plot the synthetic catalog and exposures
ICAMS::PlotCatalogToPdf(catalog = DBS78.synthetic.tumors$ground.truth.catalog,
                        file = file.path(tempdir(), "DBS78.synthetic.catalog.pdf"))
ICAMSxtra::PlotExposureToPdf(exposure = DBS78.synthetic.tumors$ground.truth.exposures,
                             file = file.path(tempdir(), "DBS78.synthetic.exposures.pdf"),
                             cex.xaxis = 0.7)
}
