% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateSyntheticTumorsFromSigParams.R
\name{GenerateSyntheticTumorsFromSigParams}
\alias{GenerateSyntheticTumorsFromSigParams}
\title{Generate synthetic tumors based on signature parameters in one or more cancer types}
\usage{
GenerateSyntheticTumorsFromSigParams(
  seed,
  dir,
  cancer.types,
  samples.per.cancer.type,
  input.sigs,
  sig.params,
  distribution = NULL,
  sample.prefix.name = "SP.Syn.",
  tumor.marker.name = NULL,
  overwrite = TRUE,
  verbose = 0
)
}
\arguments{
\item{seed}{A random seed to use.}

\item{dir}{The directory in which to put the output; will be created if
necessary.}

\item{cancer.types}{A vector of character strings denoting different cancer
types. See \code{PCAWG7::CancerTypes()} for example.}

\item{samples.per.cancer.type}{Number of synthetic tumors to create for each
cancer type. If it is \strong{one} number, then generate the \strong{same}
number of synthetic tumors for each \code{cancer.types}. Or if it is a
\strong{vector} of numbers, then generate synthetic tumors for each
\code{cancer.type} accordingly to the number specified in the vector. The length
and order of \code{samples.per.cancer.type} should match that in \code{cancer.types}.}

\item{input.sigs}{A matrix of signatures.}

\item{sig.params}{A list of empirical signature parameters generated using
real exposures for each \code{cancer.types}. This list should have names
that match \code{cancer.types}. Users can use
\code{GenerateListOfSigParams} to generate their own signature parameters.}

\item{distribution}{Probability distribution used to generate synthetic
exposures due to active mutational signatures. Can be \code{neg.binom}
which stands for negative binomial distribution. If \code{NULL} (Default),
then this function uses log normal distribution with base 10.}

\item{sample.prefix.name}{Prefix name to add to the synthetic tumors.}

\item{tumor.marker.name}{Tumor marker name to add to the synthetic tumors.
E.g. "MSI", "POLE".}

\item{overwrite}{If TRUE, overwrite existing directories and files.}

\item{verbose}{If > 0 cat various messages.}
}
\value{
A list of three elements that comprise the
synthetic data: \enumerate{
 \item \code{ground.truth.catalog}: Spectra catalog with rows denoting mutation
 types and columns denoting sample names.
 \item \code{ground.truth.signatures}: Signatures active
 in \code{ground.truth.catalog}.
 \item \code{ground.truth.exposures}: Exposures of \code{ground.truth.signatures}
 in \code{ground.truth.catalog}.
}
}
\description{
Generate synthetic tumors based on signature parameters in one or more cancer types
}
\note{
See also \code{GenerateSyntheticTumors} which uses real exposure matrix
instead of list of signature parameters to generate synthetic tumor.
}
\examples{

# Generate synthetic tumors for DBS78 using log normal distribution
input.sigs.DBS78 <- cosmicsig::COSMIC_v3.2$signature$GRCh37$DBS78
real.exposures.DBS78 <- PCAWG7::exposure$PCAWG$DBS78
cancer.types <- PCAWG7::CancerTypes()[1:5]
sig.params <- SynSigGen::signature.params$DBS78
DBS78.sig.params <-
  GenerateListOfSigParams(real.exposures = real.exposures.DBS78,
                          cancer.types = cancer.types,
                          sig.params = sig.params)

DBS78.synthetic.tumors <-
  GenerateSyntheticTumorsFromSigParams(seed = 191906,
                                       dir = file.path(tempdir(), "DBS78.synthetic.tumors"),
                                       cancer.types = cancer.types,
                                       samples.per.cancer.type = 30,
                                       input.sigs = input.sigs.DBS78,
                                       sig.params = DBS78.sig.params,
                                       sample.prefix.name = "SP.Syn.")

# Generate synthetic tumors for Indel (ID) using negative binomial distribution
input.sigs.ID <- cosmicsig::COSMIC_v3.2$signature$GRCh37$ID
real.exposures.ID <- PCAWG7::exposure$PCAWG$ID
cancer.types <- PCAWG7::CancerTypes()[1:5]
sig.params <- SynSigGen::signature.params$ID
ID.sig.params <-
  GenerateListOfSigParams(real.exposures = real.exposures.ID,
                          cancer.types = cancer.types,
                          distribution = "neg.binom",
                          sig.params = sig.params)

ID.synthetic.tumors <-
  GenerateSyntheticTumorsFromSigParams(seed = 191906,
                                       dir = file.path(tempdir(), "ID.synthetic.tumors"),
                                       cancer.types = cancer.types,
                                       samples.per.cancer.type = 30,
                                       input.sigs = input.sigs.ID,
                                       sig.params = ID.sig.params,
                                       distribution = "neg.binom",
                                       sample.prefix.name = "SP.Syn.")

# Plot the synthetic catalog and exposures
ICAMS::PlotCatalogToPdf(catalog = DBS78.synthetic.tumors$ground.truth.catalog,
                        file = file.path(tempdir(), "DBS78.synthetic.catalog.pdf"))
mSigTools::plot_exposure_to_pdf(exposure = DBS78.synthetic.tumors$ground.truth.exposures,
                               file = file.path(tempdir(), "DBS78.synthetic.exposures.pdf"),
                               cex.xaxis = 0.7)
}
