% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSynData.R
\name{GetNonZeroNoisySample}
\alias{GetNonZeroNoisySample}
\title{Add noise to one sample until the final spectrum has mutation > 0}
\usage{
GetNonZeroNoisySample(
  one.sample,
  input.exposure,
  signatures,
  n.binom.size = 100
)
}
\arguments{
\item{one.sample}{Name of one sample to add noise.}

\item{input.exposure}{The exposures to which to add noise; a numeric matrix
or data frame in which the rows are signatures and the columns are
samples. Each cell indicates the number of mutations due to a particular
signature in a particular sample.}

\item{signatures}{The signatures in the exposure; the column names
of \code{signatures} have to include all row names in
\code{input.exposure}; can be an \code{\link[ICAMS]{ICAMS}}
catalog or a numerical matrix or data frame.}

\item{n.binom.size}{If non \code{NULL}, use negative binomial noise
with this size parameter; see \code{\link[stats]{NegBinomial}}.
If \code{NULL}, use Poisson noise.}
}
\description{
Add noise to one sample until the final spectrum has mutation > 0
}
