% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSynData.R
\name{GetSynSigParamsFromExposures}
\alias{GetSynSigParamsFromExposures}
\title{Empirical estimates of key parameters describing exposures due to signatures.}
\usage{
GetSynSigParamsFromExposures(
  exposures,
  verbose = 0,
  distribution = NULL,
  cancer.type = NULL,
  sig.params = NULL
)
}
\arguments{
\item{exposures}{A matrix in which each column is a sample and each row is a mutation
signature, with each element being the "exposure",
i.e. mutation count attributed to a
(sample, signature) pair.}

\item{verbose}{If > 0 cat various messages.}

\item{distribution}{Probability distribution used to fit exposures due to one
mutational signature. Can be \code{neg.binom} which stands for negative
binomial distribution. If \code{NULL} (Default), then this function uses
log normal distribution with base 10.}

\item{cancer.type}{Optional argument specifying the cancer type of the
samples being analyzed.}

\item{sig.params}{Empirical signature parameters generated using real
exposures irrespective of their cancer types. If there
is only one tumor having a signature in a cancer type in \code{exposures},
we cannot fit the \code{distribution} to only one data point. Instead, we
will use the empirical parameter from \code{sig.params}.
Users can use \code{SynSigGen:::GetSynSigParamsFromExposuresOld} to generate
their own signature parameters. If \code{NULL}(default), this function uses the
PCAWG7 empirical signature parameters. See \code{signature.params} for more details.}
}
\value{
\itemize{
\item For log normal distribution,
a data frame with one column for
each of a subset of the input signatures
and the following rows
\describe{
\item{prob}{The proportion of tumors with the signature.}
\item{mean}{The mean(log_10(number of mutations)).}
\item{stdev}{The stdev(log_10(number of mutations)).}
}
Signatures not present in
\code{exposures} or present only in a single tumor in
\code{exposures} are removed.
\item For negative binomial distribution,
a data frame with one column for
each of a subset of the input signatures
and the following rows
\describe{
\item{prob}{The proportion of tumors with the signature.}
\item{size}{Dispersion parameter.}
\item{mu}{Mean.}
}
}
}
\description{
Empirical estimates of key parameters describing exposures due to signatures.
}
