% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ID.MMR.params}
\alias{ID.MMR.params}
\title{ID signature parameters from MMR (mismatch repair) deficient tumors}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 3 rows and 2 columns.
}
\usage{
ID.MMR.params
}
\description{
ID signature parameters from MMR (mismatch repair) deficient tumors
}
\note{
These parameters were generated using the MMR deficient tumors in
  PCAWG7 platinum ID exposures irrespective of their cancer types.
}
\keyword{datasets}
