% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateFromReal_Nat.R
\name{ManyTypes2700}
\alias{ManyTypes2700}
\title{Create a specific synthetic data set of 2,700 tumors.}
\usage{
ManyTypes2700(
  seed = 191906,
  regress.dir = "data-raw/long.test.regression.data/syn.many.types/",
  top.level.dir = "../Many.types.191906",
  unlink = FALSE
)
}
\arguments{
\item{seed}{A random seed to use.}

\item{regress.dir}{If not \code{NULL}, compare the result to
the contents of this directory with a \code{diff}.}

\item{top.level.dir}{The directory in which to put the output; will
be created if necessary.}

\item{unlink}{If \code{TRUE} and \code{!is.null(regress.dir)}, then
unlink the result directory if there are no differences.}
}
\description{
Data set generated by this function can be found at Synapse with Synapse ID:
\href{https://www.synapse.org/#!Synapse:syn18500213}{syn18500213}.
}
