% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diff4SynDataSets.R
\name{NewDiff4SynDataSets}
\alias{NewDiff4SynDataSets}
\title{Diff two directories or files.}
\usage{
NewDiff4SynDataSets(newdir, regressdirname, unlink, verbose = FALSE,
  long.diff = FALSE)
}
\arguments{
\item{newdir}{the path of \code{dir2} for a folder to be recursively compared with
\code{dir1}; it can also be the path of a single file \code{file2} to diff with
\code{file1}.}

\item{regressdirname}{the path of \code{dir2} for a folder to be recursively compared with
\code{dir1}; it can also be the path of a single file \code{file2} to diff with
\code{file1}.}

\item{unlink}{if \code{TRUE} unlink \code{newdir}, but do not unlink
if there are diffs.}

\item{verbose}{Whether to display additional R messages.}

\item{long.diff}{If \code{TRUE}, invoke "diff -r" (detailed text information
even if the two files/folders are the same); if \code{FALSE}, invoke "diff -rq"
(detailed text information only if two files/folders are different).
(Default: FALSE)}
}
\value{
The output of the diff command.
}
\description{
Diff two directories or files.
}
