% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateFromReal_Nat.R
\name{PancAdenoCA1000}
\alias{PancAdenoCA1000}
\title{Create 1000 synthetic pancreatic adenocarcinoma spectra.}
\usage{
PancAdenoCA1000(
  seed = 191907,
  regress.dir = "data-raw/long.test.regression.data/syn.pancreas/",
  num.syn.tumors = 1000,
  top.level.dir = "../Pan-AdenoCA",
  unlink = FALSE
)
}
\arguments{
\item{seed}{A random seed to use.}

\item{regress.dir}{If not \code{NULL}, compare the result to
the contents of this directory with a \code{diff}.}

\item{num.syn.tumors}{Generate this number of synthetic tumors.}

\item{top.level.dir}{The directory in which to put the output; will
be created if necessary.}

\item{unlink}{If \code{TRUE} and \code{!is.null(regress.dir)}, then
unlink the result directory if there are no differences.}
}
\description{
This function generates synthetic tumor spectra with mutational
signature prevalence and mutation load similar to pancreatic
adenocarcinoma in PCAWG cohort.
}
\details{
This function replaces \code{data-raw/Create.pancreas.Rmd} in GitHub
repository \code{steverozen/SynSig}. With default arguments, this
function generates the same results as \code{data-raw/Create.pancreas.Rmd}.

Data set generated by this function can be found at Synapse with Synapse ID:
\href{https://www.synapse.org/#!Synapse:syn18500212}{syn18500212}.
}
