% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewSets2019Aug.R
\name{ParametersSALike}
\alias{ParametersSALike}
\title{Parameters for generating data with exposures resembling SignatureAnalyzer exposures.}
\usage{
ParametersSALike()
}
\description{
Parameters for generating data with exposures resembling SignatureAnalyzer exposures.
}
\keyword{internal}
