% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateRandomExpRandomSigs.R
\name{ParametersSPLike}
\alias{ParametersSPLike}
\title{Parameters for generating data with exposures resembling SigProfiler exposures.}
\usage{
ParametersSPLike()
}
\description{
Parameters for generating data with exposures resembling SigProfiler exposures.
}
\keyword{internal}
