% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateFromReal_Nat.R
\name{RCCOvary1000}
\alias{RCCOvary1000}
\title{Create synthetic spectra based on renal cell carcinoma and ovarian adenocarcinoma}
\usage{
RCCOvary1000(
  seed = 191905,
  unlink = FALSE,
  regress.dir = NULL,
  top.level.dir = "tmp.3.5.40.RCC.and.ovary"
)
}
\arguments{
\item{seed}{A random seed to use.}

\item{unlink}{The directory in which to put the output; will
be created if necessary.}

\item{regress.dir}{If not \code{NULL}, compare the result to
the contents of this directory with a \code{diff}.}

\item{top.level.dir}{The directory in which to put the output; will
be created if necessary.}
}
\description{
Creates spectra dataset consists of 500 synthetic renal cell carcinoma (RCC)
with high prevalence and mutation load from SBS5 and SBS40 signatures,
and 500 synthetic ovarian adenocarcinoma with high prevalence and
mutation load from SBS3. This dataset challenges the computational approaches
as these three signatures are "flat" signatures hard to be extracted accurately.
}
\details{
This function Replaces the first part of \code{data-raw/Create.3.5.40.Rmd} in
GitHub repository \code{steverozen/SynSig}. With default arguments, this
function generates the same results as the first part of
\code{data-raw/Create.3.5.40.Rmd}.

The second half of \code{data-raw/Create.3.5.40.Rmd} in \code{steverozen/SynSig}
is replaced by \code{\link{Create.3.5.40.Abstract}}.

Data set generated by this function can be found at Synapse with Synapse ID:
\href{https://www.synapse.org/#!Synapse:syn18500214}{syn18500214}.
}
