% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SBS1SBS5datasetNames}
\alias{SBS1SBS5datasetNames}
\title{Names of datasets generated by \code{\link{CreateSBS1SBS5CorrelatedSyntheticData}}.}
\format{
Character vector specifying names of datasets.
}
\usage{
SBS1SBS5datasetNames
}
\description{
Names of datasets generated by \code{\link{CreateSBS1SBS5CorrelatedSyntheticData}}.
}
\note{
\code{S} is the initial of "slope", which refers to
\strong{SBS1-SBS5 Exposure Ratio}. \code{Rsq} is the abbreviation for
"Pearson's R squared", which refers to \strong{SBS1-SBS5 Correlation}.
}
\keyword{datasets}
