% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SBS1SBS5parameter}
\alias{SBS1SBS5parameter}
\title{Parameters used to generate synthetic spectra with correlated SBS1 and SBS5 exposures.}
\format{
A data.frame with parameters for generating the synthetic data.
The row names of the data.frame refers to the names of datasets, equals to
\code{\link{SBS1SBS5datasetNames}}.
}
\usage{
SBS1SBS5parameter
}
\description{
Parameters used to generate synthetic spectra with correlated SBS1 and SBS5 exposures.
}
\keyword{datasets}
