% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sig.params}
\alias{sig.params}
\title{Empirical signature parameters using negative binomial distribution}
\format{
An object of class \code{list} of length 3.
}
\usage{
sig.params
}
\description{
Empirical signature parameters using negative binomial distribution
}
\note{
These parameters were generated using all tumors in PCAWG7 exposure
  irrespective of their cancer types.
}
\keyword{datasets}
