% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{signature.params}
\alias{signature.params}
\title{Empirical signature parameters of PCAWG7 platinum tumor exposures using
negative binomial distribution}
\format{
An object of class \code{list} of length 3.
}
\usage{
signature.params
}
\description{
Empirical signature parameters of PCAWG7 platinum tumor exposures using
negative binomial distribution
}
\note{
These parameters were generated using all platinum tumors in PCAWG7
  exposures irrespective of their cancer types.
}
\keyword{datasets}
