% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSynData.R
\name{AddNoise}
\alias{AddNoise}
\title{Exposures and spectra with Poisson or negative binomial noise in exposures.}
\usage{
AddNoise(input.exposure, signatures, n.binom.size = NULL, cp.factor = NULL)
}
\arguments{
\item{input.exposure}{The exposures to which to add noise; a numeric matrix
or data frame in which the rows are signatures and the columns are
samples. Each cell indicates the number of mutations due to a particular
signature in a particular sample.}

\item{signatures}{The signatures in the exposure; the column names
of \code{signatures} have to include all row names in
\code{input.exposure}; can be an \code{\link[ICAMS]{ICAMS}}
catalog or a numerical matrix or data frame.}

\item{n.binom.size}{If non \code{NULL}, use negative binomial noise
with this size parameter; see \code{\link[stats]{NegBinomial}}.
If \code{NULL}, use Poisson noise.}

\item{cp.factor}{Concentration parameter factor. If non \code{NULL}, use
Dirichlet-multinomial distribution to add noise. When adding noise, the
parameter alpha of the Dirichlet multinomial distribution is equal to the
probabilities derived from partial spectra due to signature times
\code{cp.factor}. See \code{\link[MGLM]{rdirmn}} for more details about the
parameterization. If \code{NULL}, use Poisson noise.}
}
\value{
A list with the elements \describe{
\item{expsoures}{The numbers of mutations due to each signature
   after adding noise}
\item{spectra}{The spectra based on the noisy signature exposures.}
}
}
\description{
Exposures and spectra with Poisson or negative binomial noise in exposures.
}
