% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateFromReal.R
\name{BladderSkin1000}
\alias{BladderSkin1000}
\title{Generate synthetic data sets modeled on bladder TCC and skin melanoma.}
\usage{
BladderSkin1000(seed = 191906, regress = FALSE)
}
\arguments{
\item{seed}{A random seed to use.}

\item{regress}{Whether to compare the result with
local copy of dataset using a \code{diff}.}
}
\description{
Creates spectra dataset consists of 500 synthetic bladder
transitional cell carcinoma with high prevalence and mutation
load from SBS2, and 500 synthetic skin melanoma
with high prevalence and mutation load from SBS7a and SBS7b. This
dataset challenges the computational approaches as SBS2 has a similar
pattern to the mixture of SBS7a and SBS7b, thus the existence of these
signatures may interfere computational approaches from accurately
extracting these signatures.
}
\details{
This function replaces the first part of \code{data-raw/Create.2.7a.7b.Rmd}
in GitHub repository \code{steverozen/SynSig}. With default arguments, this
function generates the same results as the first part of
\code{data-raw/Create.2.7a.7b.Rmd}.

#' The second half of \code{data-raw/Create.2.7a.7b.Rmd}
is replaced by \code{\link{Create.2.7a.7b.Abstract}}.

Data set generated by this function can be found at Synapse with Synapse ID:
\href{https://www.synapse.org/#!Synapse:syn18500217}{syn18500217}.
}
