% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateFromReal.R
\name{BladderSkin1000}
\alias{BladderSkin1000}
\title{Generate synthetic data sets modeled on bladder TCC and skin melanoma.}
\usage{
BladderSkin1000(seed = 191906, regress = FALSE)
}
\description{
Replaces the first part of \code{data-raw/Create.2.7a.7b.Rmd}.
}
\details{
#' The second half of \code{data-raw/Create.2.7a.7b.Rmd}
is replaced by
\code{\link{Create.2.7a.7b.Abstract}}.
}
\keyword{internal}
