% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create.syn.many.types.R
\name{BladderSkin1000}
\alias{BladderSkin1000}
\title{Generate synthetic data sets modeled on bladder TCC and skin melanoma.}
\usage{
BladderSkin1000(seed = 191906, regress = FALSE)
}
\description{
Generate synthetic data sets modeled on bladder TCC and skin melanoma.
}
\keyword{internal}
