% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create.abstract.R
\name{Create.2.7a.7b.Abstract}
\alias{Create.2.7a.7b.Abstract}
\title{This function supersedes the second part of \code{data-raw/Create.2.7a.7b.Rmd}.}
\usage{
Create.2.7a.7b.Abstract(seed = 55, overwrite = TRUE,
  regress.dir = "data-raw/long.test.regression.data/syn.2.7a.7b.abst/")
}
\description{
With default arguments, this function generates the same results as
the second half of \code{data-raw/Create.2.7a.7b.Rmd}.
}
\keyword{internal}
