% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create.syn.many.types.R
\name{Create.syn.many.types}
\alias{Create.syn.many.types}
\title{No longer used, replaced by \code{\link{ManyTypes2700}}.}
\usage{
Create.syn.many.types(regress = FALSE, seed = NULL, unlink = FALSE)
}
\description{
No longer used, replaced by \code{\link{ManyTypes2700}}.
}
\keyword{internal}
