% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create.syn.many.types.R
\name{CreateFromReal}
\alias{CreateFromReal}
\title{Create a specific synthetic data set based on real exposures in one or more cancer types.}
\usage{
CreateFromReal(seed, enclosing.dir, num.syn.tumors, cancer.types,
  data.suite.name, sa.exp = SynSigGen::sa.all.real.exposures,
  sp.exp = SynSigGen::sp.all.real.exposures, overwrite = TRUE,
  regress.dir = NULL, unlink = FALSE, bladder.regress.hack = FALSE)
}
\description{
Create a specific synthetic data set based on real exposures in one or more cancer types.
}
\keyword{internal}
