% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateFromReal.R
\name{CreateFromReal}
\alias{CreateFromReal}
\title{Create a specific synthetic data set based on real exposures in one or more cancer types.}
\usage{
CreateFromReal(seed, top.level.dir = NULL, enclosing.dir = NULL,
  num.syn.tumors, cancer.types, data.suite.name = NULL,
  sa.exp = SynSigGen::sa.all.real.exposures,
  sp.exp = SynSigGen::sp.all.real.exposures, overwrite = TRUE,
  regress.dir = NULL, unlink = FALSE, verbose = FALSE,
  bladder.regress.hack = FALSE)
}
\arguments{
\item{seed}{A random seed to use.}

\item{top.level.dir}{The directory in which to put the output; will
be created if necessary.}

\item{enclosing.dir}{Deprecated; create the output in a subdirectory of this directory.}

\item{num.syn.tumors}{The number of tumors to create \strong{for each cancer
type} in \code{cancer.types}.}

\item{cancer.types}{Search \code{sa.exp} and \code{sp.exp}
for exposures from tumors matching
these strings. Each string should identify one tumor type, for
some definition of tumor type. Probably the tumors in each type
should be non-overlapping, but the code does not enforce this and
does not care.}

\item{data.suite.name}{Deprecated; the directory created will be
\code{file.path(enclosing.dir, paste0(data.suite.name, ".", seed))}.}

\item{sa.exp}{A matrix of exposures; this function will use the
columns with column names beginning \code{paste0(cancer.type, "::")}.}

\item{sp.exp}{A matrix of exposures; this function will use the
columns with column names beginning \code{paste0(cancer.type, "::")}.}

\item{overwrite}{If TRUE, overwrite existing directories and files.}

\item{regress.dir}{If not \code{NULL}, compare the result to
the contents of this directory with a \code{diff}.}

\item{unlink}{If \code{TRUE} and \code{!is.null(regress.dir)}, then
unlink the result directory if there are no differences.}

\item{verbose}{If \code{TRUE} print various informative messages.}

\item{bladder.regress.hack}{Set this to \code{TRUE} to handle
mixed "all" and "no hyper" signature sets for the
regression test for \code{\link{BladderSkin1000}}.}
}
\description{
Create a full SignatureAnalyzer / SigProfiler test data set for a
set of various tumor types.
}
