% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateRandomExpRandomSigs.R
\name{GenerateAllRandomSP}
\alias{GenerateAllRandomSP}
\title{Generate a suite of data with random signatures and exposures.}
\usage{

  GenerateAllRandomSP(top.level.dir = "../SP.like.rand.exp.rand.sigs.2019.10.28",
  overwrite = TRUE, verbose = TRUE, num.replicates = 10,
  unlink = FALSE)
}
\description{
Generate a suite of data with random signatures and exposures.
}
\keyword{(exposures).}
\keyword{Base}
\keyword{SigProfiler}
\keyword{attributions}
\keyword{internal}
\keyword{on}
\keyword{parameters}
