% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateNoisyTumors.R
\name{GenerateNoisyTumors}
\alias{GenerateNoisyTumors}
\title{Generate noisy tumors from available exposures}
\usage{
GenerateNoisyTumors(
  seed,
  dir,
  input.exposure,
  signatures,
  n.binom.size = NULL,
  overwrite = TRUE
)
}
\arguments{
\item{seed}{A random seed to use.}

\item{dir}{The directory in which to put the output; will be created if
necessary.}

\item{input.exposure}{A matrix of exposures.}

\item{signatures}{A matrix of signatures.}

\item{n.binom.size}{If non \code{NULL}, use negative binomial noise
with this size parameter; see \code{\link[stats]{NegBinomial}}.
If \code{NULL}, then use Poisson distribution to do the resampling.}

\item{overwrite}{If TRUE, overwrite existing directories and files.}
}
\value{
A list with the elements \describe{
\item{expsoures}{The numbers of mutations due to each signature
   after adding noise}
\item{spectra}{The spectra based on the noisy signature exposures.}
}
}
\description{
Generate noisy tumors from available exposures
}
\examples{

# Generate synthetic tumors for Indel (ID) using negative binomial distribution
input.sigs.ID <- PCAWG7::signature$genome$ID
real.exposures.ID <- PCAWG7::exposure$PCAWG$ID
cancer.types <- PCAWG7::CancerTypes()[1:5]
ID.synthetic.tumors <-
  GenerateSyntheticTumors(seed = 191906,
                          dir = file.path(tempdir(), "ID.synthetic.tumors"),
                          cancer.types = cancer.types,
                          samples.per.cancer.type = 30,
                          input.sigs = input.sigs.ID,
                          real.exposures = real.exposures.ID,
                          distribution = "neg.binom",
                          sample.prefix.name = "SP.Syn."
  )

# Add noise to the exposures
ID.noisy.tumors <-
  GenerateNoisyTumors(seed = 892513,
                      dir = file.path(tempdir(), "ID.noisy.tumors"),
                      input.exposure = ID.synthetic.tumors$ground.truth.exposures,
                      signatures = ID.synthetic.tumors$ground.truth.signatures,
                      n.binom.size = 1)

# Plot the synthetic and noisy catalog and exposures
ICAMS::PlotCatalogToPdf(catalog = ID.synthetic.tumors$ground.truth.catalog,
                        file = file.path(tempdir(), "ID.synthetic.catalog.pdf"))
ICAMSxtra::PlotExposureToPdf(exposure = ID.synthetic.tumors$ground.truth.exposures,
                             file = file.path(tempdir(), "ID.synthetic.exposures.pdf"),
                             cex.xaxis = 0.7)
ICAMS::PlotCatalogToPdf(catalog = ID.noisy.tumors$spectra,
                        file = file.path(tempdir(), "ID.noisy.catalog.pdf"))
ICAMSxtra::PlotExposureToPdf(exposure = ID.noisy.tumors$exposures,
                             file = file.path(tempdir(), "ID.noisy.exposures.pdf"),
                             cex.xaxis = 0.7)

}
