% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateSyntheticTumors.R
\name{GenerateSyntheticTumors}
\alias{GenerateSyntheticTumors}
\title{Generate synthetic tumors based on real exposures in one or more cancer types}
\usage{
GenerateSyntheticTumors(
  seed,
  dir,
  cancer.types,
  samples.per.cancer.type,
  input.sigs,
  real.exposures,
  distribution = NULL,
  sample.prefix.name = "SP.Syn.",
  tumor.marker.name = NULL,
  overwrite = TRUE,
  verbose = 0,
  sig.params = NULL
)
}
\arguments{
\item{seed}{A random seed to use.}

\item{dir}{The directory in which to put the output; will be created if
necessary.}

\item{cancer.types}{A vector of character strings denoting different cancer
types. This function will search \code{real.exposures} for exposures from
tumors matching these strings. See \code{PCAWG7::CancerTypes()} for example.}

\item{samples.per.cancer.type}{Number of synthetic tumors to create for each
cancer type. If it is \strong{one} number, then generate the \strong{same}
number of synthetic tumors for each \code{cancer.types}. Or if it is a
\strong{vector} of numbers, then generate synthetic tumors for each
\code{cancer.type} accordingly to the number specified in the vector. The length
and order of \code{samples.per.cancer.type} should match that in \code{cancer.types}.}

\item{input.sigs}{A matrix of signatures.}

\item{real.exposures}{A matrix of real exposures.}

\item{distribution}{Probability distribution used to generate synthetic
exposures due to active mutational signatures. Can be \code{neg.binom}
which stands for negative binomial distribution. If \code{NULL} (Default),
then this function uses log normal distribution with base 10.}

\item{sample.prefix.name}{Prefix name to add to the synthetic tumors.}

\item{tumor.marker.name}{Tumor marker name to add to the synthetic tumors.
E.g. "MSI", "POLE".}

\item{overwrite}{If TRUE, overwrite existing directories and files.}

\item{verbose}{If > 0 cat various messages.}

\item{sig.params}{Empirical signature parameters generated using real
exposures irrespective of their cancer types. If there
is only one tumor having a signature in a cancer type in \code{real.exposures},
we cannot fit the \code{distribution} to only one data point. Instead, we
will use the empirical parameter from \code{sig.params}.
Users can use \code{SynSigGen:::GetSynSigParamsFromExposuresOld} to generate
their own signature parameters. If \code{NULL}(default), this function uses the
PCAWG7 empirical signature parameters. See \code{signature.params} for more details.}
}
\value{
A list of three elements that comprise the
synthetic data: \enumerate{
 \item \code{ground.truth.catalog}: Spectra catalog with rows denoting mutation
 types and columns denoting sample names.
 \item \code{ground.truth.signatures}: Signatures active
 in \code{ground.truth.catalog}.
 \item \code{ground.truth.exposures}: Exposures of \code{ground.truth.signatures}
 in \code{ground.truth.catalog}.
}
}
\description{
Generate synthetic tumors based on real exposures in one or more cancer types
}
\examples{

# Generate synthetic tumors for DBS78
input.sigs.DBS78 <- PCAWG7::signature$genome$DBS78
real.exposures.DBS78 <- PCAWG7::exposure$PCAWG$DBS78
cancer.types <- PCAWG7::CancerTypes()[1:5]
DBS78.synthetic.tumors <-
  GenerateSyntheticTumors(seed = 191906,
                          dir = file.path(tempdir(), "DBS78.synthetic.tumors"),
                          cancer.types = cancer.types,
                          samples.per.cancer.type = 30,
                          input.sigs = input.sigs.DBS78,
                          real.exposures = real.exposures.DBS78,
                          sample.prefix.name = "SP.Syn."
  )

# Generate synthetic tumors for Indel (ID) using negative binomial distribution
input.sigs.ID <- PCAWG7::signature$genome$ID
real.exposures.ID <- PCAWG7::exposure$PCAWG$ID
cancer.types <- PCAWG7::CancerTypes()[1:5]
ID.synthetic.tumors <-
  GenerateSyntheticTumors(seed = 191906,
                          dir = file.path(tempdir(), "ID.synthetic.tumors"),
                          cancer.types = cancer.types,
                          samples.per.cancer.type = 30,
                          input.sigs = input.sigs.ID,
                          real.exposures = real.exposures.ID,
                          distribution = "neg.binom",
                          sample.prefix.name = "SP.Syn."
  )

# Plot the synthetic catalog and exposures
ICAMS::PlotCatalogToPdf(catalog = DBS78.synthetic.tumors$ground.truth.catalog,
                        file = file.path(tempdir(), "DBS78.synthetic.catalog.pdf"))
ICAMSxtra::PlotExposureToPdf(exposure = DBS78.synthetic.tumors$ground.truth.exposures,
                             file = file.path(tempdir(), "DBS78.synthetic.exposures.pdf"),
                             cex.xaxis = 0.7)
}
