% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSynData.R
\name{GetSynSigParamsFromExposures}
\alias{GetSynSigParamsFromExposures}
\title{Empirical estimates of key parameters describing exposures due to signatures.}
\usage{
GetSynSigParamsFromExposures(exposures, verbose = 0)
}
\arguments{
\item{exposures}{A matrix in which each column is a sample and each row is a mutation
signature, with each element being the "exposure",
i.e. mutation count attributed to a
(sample, signature) pair.}

\item{verbose}{If > 0 cat various messages.}
}
\value{
A data frame with one column for
each of a subset of the input signatures
and the following rows
\enumerate{
\item the proportion of tumors with the signature
\item mean(log_10(mutations.per.Mb))
\item stdev(log_10(mutations.per.Mb))
}
Signatures not present in
 \code{exposures} or present only in a single tumor in
 \code{exposures} are removed.
}
\description{
Empirical estimates of key parameters describing exposures due to signatures.
}
