% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateFromReal.R
\name{ManyTypes2700}
\alias{ManyTypes2700}
\title{Create a specific synthetic data set of 2,700 tumors.}
\usage{
ManyTypes2700(seed = 191906, regress = FALSE)
}
\arguments{
\item{seed}{A random seed to use.}

\item{regress}{Whether to compare the result with
local copy of dataset using a \code{diff}.}
}
\description{
Data set generated by this function can be found at Synapse with Synapse ID:
\href{https://www.synapse.org/#!Synapse:syn18500213}{syn18500213}.
}
