% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateFromReal.R
\name{ManyTypes2700}
\alias{ManyTypes2700}
\title{Create a specific synthetic data set of 2,700 tumors.}
\usage{
ManyTypes2700(seed = 191906, regress = FALSE)
}
\description{
Replaces \code{\link{Create.syn.many.types}}.
}
\keyword{internal}
