% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create.syn.many.types.R
\name{PancAdenoCA1000}
\alias{PancAdenoCA1000}
\title{Replaces \code{data-raw/Create.pancrease.Rmd}.}
\usage{
PancAdenoCA1000(seed = 191907,
  regress.dir = "data-raw/long.test.regression.data/syn.pancreas/",
  num.syn.tumors = 1000)
}
\description{
Replaces \code{data-raw/Create.pancrease.Rmd}.
}
\keyword{internal}
