% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create.syn.many.types.R
\name{RCCOvary1000}
\alias{RCCOvary1000}
\title{Replaces the first part of \code{data-raw/Create.3.5.40.Rmd}.}
\usage{
RCCOvary1000(seed = 191905, regress = FALSE)
}
\description{
The second half of \code{data-raw/Create.3.5.40.Rmd}
is replaced by
\code{\link{Create.3.4.40.Abstract}}.
}
\keyword{internal}
